admin privileges and security works on the back end for add/edit/delete routes
