+++ server/api/products.js
@@ -1,7 +1,6 @@
 //api get for product info
 const router = require('express').Router()
-const {Product} = require('../db/models')
-
+const {Product, User} = require('../db/models')
 
 router.get('/', async (req, res, next) => {
   try {
@@ -22,44 +21,59 @@ router.get('/:id', async (req, res, next) => {
 })
 
 router.post('/add', (req, res, next) => {
-  try {
-    Product.create({
-      title: req.body.title,
-      description: req.body.description,
-      price: req.body.price,
-      quantity: req.body.quantity,
-      imageUrl: req.body.imageUrl
-    })
-  } catch (err) {
-    next(err)
+  if(!req.user.isAdmin){
+    res.sendStatus(403)
+  } else {
+    try {
+      Product.create({
+        title: req.body.title,
+        description: req.body.description,
+        price: req.body.price,
+        quantity: req.body.quantity,
+        imageUrl: req.body.imageUrl
+      })
+      res.sendStatus(201)
+    } catch (err) {
+      next(err)
+    }
   }
 })
 
 router.put('/:id', (req, res, next) => {
-  try {
-    Product.update({
-      title: req.body.title,
-      description: req.body.description,
-      price: req.body.price,
-      quantity: req.body.quantity,
-      imageUrl: req.body.imageUrl
-    }, {
-      where: { id: req.params.id }
-    })
-  } catch (err) {
-    next(err)
+  if (!req.user.isAdmin) {
+    res.sendStatus(403)
+  } else {
+    try {
+      Product.update({
+        title: req.body.title,
+        description: req.body.description,
+        price: req.body.price,
+        quantity: req.body.quantity,
+        imageUrl: req.body.imageUrl
+      }, {
+        where: { id: req.params.id }
+      })
+      res.sendStatus(201)
+    } catch (err) {
+      next(err)
+    }
   }
 })
 
 router.delete('/:id', (req, res, next) => {
-  try {
-    Product.destroy({
-      where: {
-        id: req.params.id
-      }
-    })
-  } catch (err) {
-    next(err)
+  if (!req.user.isAdmin) {
+    res.sendStatus(403)
+  } else {
+    try {
+      Product.destroy({
+        where: {
+          id: req.params.id
+        }
+      })
+      res.sendStatus(204)
+    } catch (err) {
+      next(err)
+    }
   }
 })
