+++ server/api/products.spec.js
@@ -5,10 +5,19 @@ const request = require('supertest')
 const db = require('../db')
 const app = require('../index')
 const Product = db.model('product')
+const User = db.model('user')
 
 describe('Product routes', () => {
   describe('/api/products/', () => {
     const testTitle = 'Title Of Product'
+    const cody = {
+      email: 'cody@puppybook.com',
+      name: 'Cody',
+      isAdmin: true
+    }
+
+    beforeEach(() => {
+    })
 
     beforeEach(() => {
       Product.create({
@@ -32,6 +41,7 @@ describe('Product routes', () => {
         quantity: 203,
         imageUrl: 'defaultImage3.jpg',
       })
+      return User.create(cody)
     })
 
     it('GET /api/products', () => {
@@ -51,6 +61,22 @@ describe('Product routes', () => {
           expect(res.body).to.be.an('object')
           expect(res.body.title).to.be.equal(testTitle + '2')
         })
+    })
+    it('/api/admin check if authenticated', async () => {
+      const res = await request(app).post('/api/admin/addproduct').expect()
+    })
+
+    it('POST /api/admin/addproduct', async () => {
+      const res = await request(app).post('/api/admin/addproduct').expect(201)
+      expect(res.body).to.be.an('array')
+      expect(res.body[0].email).to.be.equal(cody.email)
+      expect(res.body[0].isAdmin).to.be.equal(true)
+    })
+
+    it('PUT /api/admin/editproduct', async () => {
+      const res = await request(app).post('/api/admin/editproduct', { id: 1, name: 'editName' }).expect(201)
+      expect(res.body).to.be.an('array')
+      expect(res.body[0].email).to.be.equal(codysEmail)
     })
   }) // end describe('/api/users')
 }) // end describe('User routes')
